backend/requirements.elite.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
redis==5.0.1
numpy==1.24.3
pillow==10.1.0
python-multipart==0.0.6
